Addressed the initialization issue
